requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.0.1
Flask-DebugToolbar==0.9.2
Flask-Uploads==0.1.3
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Pygments==2.0.2
WTForms==2.0.2
Werkzeug==0.10.1
argparse==1.2.1
blinker==1.3
itsdangerous==0.24
lxml==3.4.2
wsgiref==0.1.2
